+++ poem-openapi-derive/src/oneof.rs
@@ -100,18 +100,18 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
             type ValueType = Self;
 
             fn schema_ref() -> #crate_name::registry::MetaSchemaRef {
-                #crate_name::registry::MetaSchemaRef::Inline(#crate_name::registry::MetaSchema {
+                #crate_name::registry::MetaSchemaRef::Inline(Box::new(#crate_name::registry::MetaSchema {
                     one_of: ::std::vec![#(<#types as #crate_name::types::Type>::schema_ref()),*],
-                    properties: ::std::vec![(#property_name, #crate_name::registry::MetaSchemaRef::Inline(#crate_name::registry::MetaSchema {
+                    properties: ::std::vec![(#property_name, #crate_name::registry::MetaSchemaRef::Inline(Box::new(#crate_name::registry::MetaSchema {
                         enum_items: ::std::vec![#(::std::convert::Into::into(#names)),*],
                         ..#crate_name::registry::MetaSchema::new("string")
-                    }))],
+                    })))],
                     discriminator: ::std::option::Option::Some(#crate_name::registry::MetaDiscriminatorObject {
                         property_name: #property_name,
                         mapping: ::std::vec![#(#mapping),*],
                     }),
                     ..#crate_name::registry::MetaSchema::new("object")
-                })
+                }))
             }
 
             fn register(registry: &mut #crate_name::registry::Registry) {