+++ poem-openapi/src/types/multipart/upload.rs
@@ -76,7 +76,7 @@ impl Type for Upload {
     };
 
     fn schema_ref() -> MetaSchemaRef {
-        MetaSchemaRef::Inline(Self::NAME.into())
+        MetaSchemaRef::Inline(Box::new(Self::NAME.into()))
     }
 
     impl_value_type!();