+++ poem-openapi/tests/multipart.rs
@@ -475,13 +475,13 @@ fn inline_field() {
     );
     assert_eq!(
         meta_inner_obj.all_of[1],
-        MetaSchemaRef::Inline(MetaSchema {
+        MetaSchemaRef::Inline(Box::new(MetaSchema {
             title: Some("Inner Obj"),
             default: Some(serde_json::json!({
                 "v": 100,
             })),
             ..MetaSchema::ANY
-        })
+        }))
     );
 
     let meta_inner_enum = schema.properties[1].1.unwrap_inline();
@@ -491,10 +491,10 @@ fn inline_field() {
     );
     assert_eq!(
         meta_inner_enum.all_of[1],
-        MetaSchemaRef::Inline(MetaSchema {
+        MetaSchemaRef::Inline(Box::new(MetaSchema {
             title: Some("Inner Enum"),
             default: Some(serde_json::json!("B")),
             ..MetaSchema::ANY
-        })
+        }))
     );
 }