+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- [[#57]] Add `prefix_path` and `tag` attributes for `#[OpenApi]`
+
+### Breaking Changes
+
+- [[#59]] `MetaSchemaRef::Inline(MetaSchema)` -> `MetaSchemaRef::Inline(Box<MetaSchema>)`
+
 ## [1.0.2] 2021-10-11
 
 - Add `write_only` and `read_only` attributes for object fields.